feat: Implementação do service SolicitacaoService. Implementação do desing pattern State para controle das transições dos status. Criação do README.md para documentação

+++ src/main/java/br/com/sergipetec/processoseletivo/repository/SolicitanteRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface SolicitanteRepository extends JpaRepository<Solicitante, Integer> {
+public interface SolicitanteRepository extends JpaRepository<Solicitante, Long> {
 }
